Add files via upload
This project is focused on the analysis of users who are using instagram paltform. User analysis is the process by which we track how users engage and interact with our digital product (software or mobile application) in an attempt to derive business insights for marketing, product & development teams.This project will answer the problem statements which will help the product team.

A) Marketing: The marketing team wants to launch some campaigns, and they need your help with the following

1.Rewarding Most Loyal Users: People who have been using the platform for the longest time. Task: Find the 5 oldest users of the Instagram from the database provided

2.Remind Inactive Users to Start Posting: By sending them promotional emails to post their 1st photo. Task: Find the users who have never posted a single photo on Instagram

3.Declaring Contest Winner: The team started a contest and the user who gets the most likes on a single photo will win the contest now they wish to declare the winner. Task: Identify the winner of the contest and provide their details to the team

4.Hashtag Researching: A partner brand wants to know, which hashtags to use in the post to reach the most people on the platform. Task: Identify and suggest the top 5 most commonly used hashtags on the platform

5.Launch AD Campaign: The team wants to know, which day would be the best day to launch ADs. Task: What day of the week do most users register on? Provide insights on when to schedule an ad campaign

B) Investor Metrics: Our investors want to know if Instagram is performing well and is not becoming redundant like Facebook, they want to assess the app on the following grounds:

1.User Engagement: Are users still as active and post on Instagram or they are making fewer posts. Task: Provide how many times does average user posts on Instagram. Also, provide the total number of photos on Instagram/total number of users
Bots & Fake Accounts: The investors want to know if the platform is crowded with fake and dummy accounts.